+++ R/tree_models_to_xml_tracelog.R
@@ -13,14 +13,9 @@
 tree_models_to_xml_tracelog <- function(
   site_models
 ) {
-  text <- NULL
-  for (site_model in site_models) {
-    id <- site_model$id
-    text <- c(text, paste0("<log idref=\"treeLikelihood.", id, "\"/>")) # nolint this is no absolute path
-    text <- c(text, paste0("<log id=\"TreeHeight.t:", id, "\" ",
-      "spec=\"beast.evolution.tree.TreeHeightLogger\" ",
-      "tree=\"@Tree.t:", id, "\"/>") # nolint this is no absolute path
-    )
-  }
-  text
+  stop(
+    "'tree_models_to_xml_tracelog' is deprecated, ",
+    "use 'tree_model_to_tracelog_xml' instead"
+  )
+
 }